Resolves PR comments + improved re-use
